Implement IR generation basics
Added a IRGenerator class to provide context for IR generation. With it
added two visitor classes (IRExprVis and IRStatementVis) generating the
bulk of IR.

Currently the code generation consists of creating a main function
which returns the double value of the lastly parsed statement (i.e. a
top-level expression, because only those possess a value). If the last
statement was a declaration, the return value is set to void. Upon
completion, he IR is then output to stdout.

The IRGenerator should optimally provide code generation to the outside
world while using the visitors internally to traverse the syntax tree
and to generate its IR.