Add function support
Add classes for function declarations and defintions (FnDecl and FnDef)
as well as a class for Function calls (CallExpr) to the AST. Add
necessary new tokens to the lexer. Implement parsing of function
declarations, defintions and calls. Enhance the AST visitors to support
the new statements. Appropriately modify the driver in the main
function.

The following syntax defintions are in EBNF (without commas).
Characters enclosed by backticks (`) are terminal symbols.

Syntax for function declarations:
    `fn` identifier `(` [identifier {`,` identifier}] `)` `;`

Syntax for function defintions:
    `fn` identifier `(` [identifier {`,` identifier}] `)` `{`
        {(declaration | top-level expression) `;`}
    `}`

Syntax for function calls:
    identifier `(` [expression {`,` expression}] `)`

Every function is expected to return a double.

Currently the parameter names in the function definition are ignored and
those of the declaration are used in the function body. It is up for
discussion whether the language should even contain declarations.