Add basic logging functionality
Created the file Log.hpp providing functions to log errors with the
option of custom return values. Added messages and helpful macros for
logging to the lexer, parser and IR generator.

The functions take an arbitrary number of arguments which are printed
out to `std::cerr`. The fuctions can thus easily display things like
floating point numbers.

The option for custom return types may for example be used when
returning from a function signaling its failure by returning a null
pointer:

    return Log::error<std::nullptr_t>("Message");

This default initializes the given type. A particular value may be
specified as a second template parameter.