Add scoping functionality to the IR generator
Create a `IdScoper` class containing scoped variables and providing
access to them. Replace the `named_values` map in `IRGenerator` by an
`IdScoper`. Adjust IR visitors to properly use the new implementation.

This change allows the use of scopes with local variables. There
currently is no way to really take advantage of it, but it is intended
for future use. This change now allows the redeclaration of variables,
discarding the old value.