Implement scopes
Add a `ScopeExpr` class to the AST. Add parsing and code generation for
scopes. Change `IdScoper` to not return a reference when accessing a
variable normally and add a function to get a reference to a value in
the current scope. Appropriately modify the code requiring those
methods. The info visitor is left as a stub.

Returning references only to values in the current scope ensures, that
new variables can only be declared in the current scope. To declare a
value in the current scope, the new `from_current_scope` function can be
used.