Use `ScopeExpr`s as function bodies
Chang the `FnDef` class to use a scope expression as its body. Change
`get_body` to get the body of the member `ScopeExpr` and provide
`get_body_scope` to directly access the `ScopeExpr` object. Accordingly
change the parser and IR generator, which now uses the recently added
`gen_scope` function for IR creation.