Add 'less than' comparison operator
Add the Token `Tok::CMP_LT` for `<` to the Lexer. Add it to the operator
precedence list of the parser and implement IR generation for it.

The comparison returns a boolean, which currently cannot be used in the
language, but is intended for future use in conditionals. A working type
system will also make this more useful.