Add if/else expressions
Add the Tokens `Tok::IF` and `Tok::ELSE` to the lexer. Implement an AST
class `IfExpr` and provide parsing and IR generation for it. Introduce a
corresponding stub method in the info visitor.

The expressions follow the simple syntax

`if` condition
    scope_expression
`else`
    scope_expression

The else clause cannot be omitted, as the if statement is an expression
and needs to return something. This will be fixed when a proper type
system is in existence.