Add custom I/O methods to the lexer
Add the self explanatory private methods `get_char` and `peek_char` to
`Lexer`. Replace all calls to C-style I/O from the lexer with the new
functions. Replace the uses of the `EOF` macro with a constexpr acquired
from `std::char_traits`.

The new peek method allows for more intuitive I/O ridding the code of
the abomination that is ungetting a character.