Add line/column display to error messages
Add a struct called `CodeLocation` for saving the current position in a
file. Add a member of the struct type to the Lexer and let the input
functions continuously update it. Implement a specialization of the
error logging functions to output the code location in the format
"<line>:<column>: " before displaying the actual error message. Modify
the error macros in the lexer and parser to use the new functions.